app/requirements.txt
psutil==6.0.0
PyYAML==6.0.2
requests==2.32.3
typer==0.12.3
python-dateutil==2.9.0.post0
